==> 00_mapping_id.ipynb <==
import pandas as pd
import requests
import os

url = "https://raw.githubusercontent.com/theFPLkiwi/theFPLkiwi/7ce79ee2c7ed8b96863ccbd1e9bbe36f80120232/2021_IDs.csv"
res = requests.get(url, allow_redirects=True)

save_path = os.getcwd() + "/data"
file_name = 'mapping_id'
completeName = os.path.join(save_path , file_name + ".csv")

with open(completeName,'wb') as file:
    file.write(res.content)
