==> 2_FPL_clean_data.ipynb <==
import numpy as np
import pandas as pd
import requests
from datetime import datetime
import json

import os
import glob

#Set Display options
pd.set_option('display.max_rows', 50)
pd.set_option('display.max_columns', 50)
pd.set_option('display.width' , 1000)

# Variables
season = "2021-22"

source_path = os.getcwd()

raw_data_path = os.path.join(source_path, "data" , season , "Raw_data/")
cleaned_data_path =  os.path.join(source_path, "data" , season , "Cleaned_data/")


all_players = pd.read_csv (os.path.join(raw_data_path, "all_players.csv"), parse_dates= ["news_added"])

all_teams = pd.read_csv (os.path.join(raw_data_path,"all_teams.csv"))

all_events = pd.read_csv (os.path.join(raw_data_path,"all_events.csv"), parse_dates= ["deadline_time"])

player_types = pd.read_csv(os.path.join(raw_data_path,"player_types.csv"))

game_phases = pd.read_csv(os.path.join(raw_data_path,"game_phases.csv"))

understat_player = pd.read_csv(os.path.join(raw_data_path,"understat/players/understat_player.csv"))

player_season_history = pd.read_csv(os.path.join(raw_data_path,"player_season_history.csv"))

player_gameweek_history = pd.read_csv(os.path.join(raw_data_path,"player_gameweek_history.csv"))

player_future_fixture = pd.read_csv(os.path.join(raw_data_path,"player_future_fixture.csv"), parse_dates = ["kickoff_time"])

import os
import glob
import pandas as pd

open_path = os.path.join(os.getcwd(),"data",season,"Raw_data/understat/teams")
os.chdir(open_path)
# os.chdir(r'C:/Users/danie/OneDrive/Documents/GitHub/FPLAnlysis/data/2020-21/Raw_data/understat/teams/')

extension = 'csv'
all_filenames = [i for i in glob.glob('*.{}'.format(extension))]

#combine all files in the list
combined_csv = pd.concat([pd.read_csv(f) for f in all_filenames ])
#export to csv
combined_csv.to_csv( "understat_teams.csv", index=False, encoding='utf-8-sig')
combined_csv.head()

## Rename columns

player_gameweek_history.head()

player_gameweek_history.rename(
    columns = {
        "element" : "player_id" , 
        "opponent_team" : "opponent_team_id" , 
        "fixture" : "fixture_id" 
        }, 
        inplace = True)

all_players.rename(
    columns = {
        "code" : "player_code", 
        "element_type" : "player_type_id" , 
        "minutes" : "total_minutes" , "id" : 
        "player_id" , "team" : "team_id" , 
        "points_per_game" : "PPG"
        }, 
        inplace=True)

all_teams.rename(
    columns = {
        "id" : "team_id" , 
        "code" : "team_code" , 
        "name" : "team_name" , 
        "short_name" : "team_short_name" , 
        "strength" : "team_strength" , 
        "form" : "team_form"
        }, 
        inplace = True)

all_events.rename(
    columns = {
        "id" : "gameweek_id"
        }, 
        inplace = True)

player_types.rename(
    columns = {
        "id" : "player_type_id" , 
        "singular_name" : "position_name" , 
        "singular_name_short" : "position_name_short"
        }, 
        inplace = True)

game_phases.rename(
    columns = {
        "id" : "phase_id"
        }, 
        inplace = True)

player_season_history.rename(
    columns = {
        "element_code" : "player_code"
        }, 
        inplace = True)

player_future_fixture.rename(
    columns = {
        "id" : "fixture_id" , 
        "team_h" : "home_team_id" , 
        "team_a" : "away_team_id" , 
        "events" : "gameweek_id" , 
        "code" : "fixture_code"
        }, 
        inplace = True)
        
#player_gameweek_history.rename(columns = {} , inplace = True)


## Drop columns

all_events.drop(
    columns = [
        "deadline_time_epoch"
        ], 
        inplace = True)

player_types.drop(
    columns = [
        "plural_name" , 
        "plural_name_short"
        ], 
        inplace = True)
        
all_players.drop(
    columns = [
        "squad_number" , 
        "special" 
        #"transfers_in" , 
        #"transfers_out" , 
        #"first_name" , 
        #"second_name"
        ], 
        inplace = True)

## Change date columns

all_players['news_added'] = all_players['news_added'].dt.tz_localize(None)
all_events['deadline_time'] = all_events['deadline_time'].dt.tz_localize(None)
player_future_fixture['kickoff_time'] = player_future_fixture['kickoff_time'].dt.tz_localize(None)

player_gameweek_history['kickoff_time'] = pd.to_datetime(player_gameweek_history['kickoff_time'])
player_gameweek_history['kickoff_time'] = player_gameweek_history['kickoff_time'].dt.tz_localize(None)


## Add Columns

def opponent_team(row):
    if row['is_home'] == True:
        val = row['away_team_id']
    elif row['is_home'] == False:
        val = row['home_team_id']
    return val

def team(row):
    if row['is_home'] == False:
        val = row['away_team_id']
    elif row['is_home'] == True:
        val = row['home_team_id']
    return val


player_future_fixture['opponent_team_id'] = player_future_fixture.apply(opponent_team,axis=1)
player_future_fixture['team_id'] = player_future_fixture.apply(team,axis=1)


## Calculate New Columns

all_players.head()

all_players['PP90'] = np.divide(all_players['total_points'] , all_players['total_minutes'])*90 

all_players['PPMM'] = np.divide(all_players['PPG'] , all_players['now_cost']/10)

all_players['VAPM'] = np.divide(all_players['PPG']-2 , all_players['now_cost']/10)

all_players[['web_name' , 'player_type_id', 'total_points' , 'total_minutes' , 'PP90' , 'PPG' , 'PPMM' , 'VAPM']][all_players['total_minutes']>100].sort_values(by='VAPM' , ascending = False).head()

## Save cleaned data

all_players.to_csv (r'./data/' + season + '/Cleaned_data/all_players.csv', index = False, header=True)
all_teams.to_csv (r'./data/' + season + '/Cleaned_data/all_teams.csv', index = False, header = True)
all_events.to_csv (r'./data/' + season + '/Cleaned_data/all_events.csv', index = False, header = True)
player_types.to_csv(r'./data/' + season + '/Cleaned_data/player_types.csv', index = False, header = True)
game_phases.to_csv(r'./data/' + season + '/Cleaned_data/game_phases.csv', index = False, header = True)

player_season_history.to_csv (r'./data/' + season + '/Cleaned_data/player_season_history.csv', index = False, header=True)
player_gameweek_history.to_csv (r'./data/' + season + '/Cleaned_data/player_gameweek_history.csv', index = False, header=True)
player_future_fixture.to_csv (r'./data/' + season + '/Cleaned_data/player_future_fixture.csv', index = False, header=True)
