==> 3_FPL_merge_data.ipynb <==
import numpy as np
import pandas as pd
import requests
from datetime import datetime
import json
import os

#Set Display options
pd.set_option('display.max_rows', 50)
pd.set_option('display.max_columns', 20)
pd.set_option('display.width' , 400)

# Variables
season = "2021-22"

source_path = os.getcwd()

cleaned_data_path =  os.path.join(source_path, "data" , season , "Cleaned_data")
merged_data_path =  os.path.join(source_path, "data" , season , "Merged_data")


all_players = pd.read_csv (os.path.join(cleaned_data_path, 'all_players.csv'), parse_dates= ["news_added"])

all_teams = pd.read_csv (os.path.join(cleaned_data_path, 'all_teams.csv'))
all_events = pd.read_csv (os.path.join(cleaned_data_path, 'all_events.csv'), parse_dates= ["deadline_time"])
player_types = pd.read_csv(os.path.join(cleaned_data_path, 'player_types.csv'))
game_phases = pd.read_csv(os.path.join(cleaned_data_path, 'game_phases.csv'))

player_season_history = pd.read_csv(os.path.join(cleaned_data_path, 'player_season_history.csv'))
player_gameweek_history = pd.read_csv(os.path.join(cleaned_data_path, 'player_gameweek_history.csv') , parse_dates = ['kickoff_time'])
player_future_fixture = pd.read_csv(os.path.join(cleaned_data_path, 'player_future_fixture.csv'), parse_dates = ["kickoff_time"])


all_teams.head()

all_players.head()

player_future_fixture.head()

# Merge to all_players

df_merge = all_teams[["team_code" , "team_strength" , "team_short_name" , "strength_overall_home" , "strength_overall_away"]]
all_players = all_players.merge(df_merge, on = 'team_code')

df_merge = player_types[['player_type_id' , 'position_name_short']]
all_players = all_players.merge(df_merge , on = "player_type_id")


all_players.head()

# Merge to player_gameweek_history

for i in all_players.columns:
    print(i)

player_gameweek_history.head()

df_merge = all_players[["player_id" , "player_type_id" , "PPG" , "PP90" , "team_strength" , "team_id"]]
player_gameweek_history = player_gameweek_history.merge(df_merge , on = "player_id")

player_gameweek_history.head()

## Save to file

all_players.to_csv (os.path.join(merged_data_path,'all_players_merged.csv'), index = False, header=True)
player_gameweek_history.to_csv(os.path.join(merged_data_path, 'player_gameweek_history.csv') , index = False, header=True)
