==> 4_FPL_my_team.ipynb <==
## Import Libraries

import os
import numpy as np
import pandas as pd
import requests
from datetime import datetime
import json
import os

# Variables
season = "2021-22"


## Paths for Files



source_path = os.getcwd()

cleaned_data_path =  os.path.join(source_path, "data" , season , "Cleaned_data")
merged_data_path =  os.path.join(source_path, "data" , season , "Merged_data")
team_data_path =  os.path.join(source_path, "data" , season , "Team_data")

## Import files

all_events = pd.read_csv(os.path.join(cleaned_data_path, 'all_events.csv'), parse_dates= ["deadline_time"])
all_players = pd.read_csv(os.path.join(merged_data_path, 'all_players_merged.csv'), parse_dates= ["news_added"]) 

## Functions

#### Connect to FPL API
def get(url):
    response = requests.get(url)
    return json.loads(response.content)

## My Team

#### Get my fantasy team lineup
now = datetime.now()
my_team_id = '272984'
current_gw = all_events[all_events['deadline_time'] < now]['gameweek_id'].max()

my_team_url = 'https://fantasy.premierleague.com/api/entry/{}/event/{}/picks/'.format(my_team_id,str(current_gw))

my_team = get(my_team_url)['picks']
my_team = pd.DataFrame(my_team)


all_players.columns

#all_players = all_players[['chance_of_playing_next_round', 'chance_of_playing_this_round', 'position_id', 'ep_next','ep_this', 'player_id', 'player_news', 'news_added', 'player_price', 'selected_by_percent', 'player_status','team_id', 'web_name']]

all_players.head()

my_team

my_team = my_team.merge(all_players, how = 'left', left_on="element", right_on="player_id")
my_team.drop(["element","position"], axis = 1, inplace = True)

my_team.sort_values(by=['player_type_id' , 'selected_by_percent'])['web_name']



my_team.to_csv (os.path.join(team_data_path, 'my_team.csv'), index = False, header=True)

my_team
