==> 7_top_managers.ipynb <==
import requests , json, csv
import pandas as pd
import os

# Season
season = '2021-22'

source_path = os.getcwd()

cleaned_data_path =  os.path.join(source_path, "data" , season , "Cleaned_data")
managers_data_path =  os.path.join(source_path, "data" , season , "managers")




# Overall FPL league ID, 314 for 2019/20 season.
#overallLeageID = 300792
#overallLeageID = 314

#Visma FPL
LeageID = 1313476

#Richie Rich - 1033085
#Visma FPL - 1313476

# number of GW in 2019/20 season. Done as array to avoid calling api for blank GW 30-38.
gameWeeks = list(range(1,39))

# adds the top 10 team ID's to this array
teamIDarray = []

# number of top manager information required.
topManagerNumber = 20

url = "https://fantasy.premierleague.com/api/leagues-classic/"+str(LeageID)+"/standings/"

response = requests.get(url)
data = response.text
parsed = json.loads(data)
managerParsed = parsed['standings']['results']

# open 3 files for writing
manager_data = open(os.path.join(managers_data_path, 'top_managers.csv'), 'w', newline='', encoding="utf-8")
gw_data = open(os.path.join(managers_data_path, 'top_managers_gwInfo.csv'), 'w', newline='', encoding="utf-8")
gw_picks = open(os.path.join(managers_data_path, 'top_managers_gwPicks.csv'), 'w', newline='', encoding="utf-8")

# create the 3 csv writer objects
csvwriter1 = csv.writer(manager_data)
csvwriter2 = csv.writer(gw_data)
csvwriter3 = csv.writer(gw_picks)

#get csv of top 10 manager information and write to top_managers.csv
count = 0
for manager in managerParsed:
	if count == 0:
		header = ['rank','entry','player_name','entry_name','total']
		csvwriter1.writerow(header)
		count += 1
	
	if count <= topManagerNumber:
		csvwriter1.writerow([manager['rank'],manager['entry'],manager['player_name'],
						manager['entry_name'],manager['total']])

		count +=1
		teamIDarray.append(manager['entry'])

# for each teamID in the top 10, call the api and update both top_managers_gwInfo.csv and top_managers_gwPicks.csv
count1 = 0
count2 = 0
for teamID in teamIDarray:
	for x in gameWeeks:
		url = "https://fantasy.premierleague.com/api/entry/"+str(teamID)+"/event/"+str(x)+"/picks/"
		response = requests.get(url)
		data = response.text
		parsed = json.loads(data)

		# write data to top_managers_gwInfo.csv
		if count1 == 0:
			header = ['team_id','gw','points','bench','gw_rank','transfers','hits','total_points',
				'overall_ank','team_value','chip']

			csvwriter2.writerow(header)
			count1 += 1
		try:
			csvwriter2.writerow([teamID,x, parsed['entry_history']['points'], parsed['entry_history']['points_on_bench'],
							parsed['entry_history']['rank'], parsed['entry_history']['event_transfers'],
							parsed['entry_history']['event_transfers_cost'], parsed['entry_history']['total_points'],
							parsed['entry_history']['overall_rank'], int(parsed['entry_history']['value'])/10, parsed['active_chip']])
		except:
			continue

		# write data to top_managers_gwPicks.csv
		for i in range(len(parsed['picks'])):
			if count2 == 0:
				header = ['team_id','gw','id','position','multiplier']
				csvwriter3.writerow(header)
				count2 += 1
			csvwriter3.writerow([teamID, x, parsed['picks'][i]['element'], parsed['picks'][i]['position'], parsed['picks'][i]['multiplier']])




df1 = pd.read_csv(os.path.join(cleaned_data_path, 'all_players.csv')) 
df2 = df1[['player_id' , 'web_name']]
df2.head()

manager_data.close()
gw_data.close()
gw_picks.close()

# do some formatting on top_managers_gwPicks by adding the name of the player picked from player_idlist.csv
df = pd.read_csv(os.path.join(managers_data_path, 'top_managers_gwPicks.csv'))
df1 = pd.read_csv(os.path.join(cleaned_data_path, 'all_players.csv')) 
df1 = df1[['player_id' , 'web_name']]


merged = df.merge(df1, left_on='id', right_on='player_id')
merged.drop('id', axis=1, inplace=True)
merged.sort_values(by=['team_id', 'gw', 'position'], inplace = True)
merged.head(50)
merged.to_csv(os.path.join(managers_data_path, 'top_managers_gwPicks_cleaned.csv'),index=False)
